refactor(agents): split test_agents.py into per-agent test files
